core-mem-ring: cap the size allocated in a block
When an allocation is asked in the ring pool and it creates a new
block, it allocates 64 times the mean allocation to handle future,
expected allocations. However, if the allocation is huge, this
can go over the limit and abort.
To avoid this, the size computed from the mean of previous
allocations is capped at 100MB.

Closes: #37299
Change-Id: I333b786bb6ef3ac3cc120d04d4e8ed1528204827