httd_query: add header in ichttp_query.
add possibility to fill a header when using web services.
Sometime we need to get user, for example get restriction for filter the
schema. The user is get in the header but when using web services, the
header is not filled.

If need to have a header when use web services, we can filled field header
in ichttp_query.

Refs: #37183
Change-Id: Ic0d5ad520cedd8548e9437a4a93b2517bb2a36fc