qps: use the el-loop to wait for the background snapshotter.
Previously, we waited for the snapshotter in a dedicated thread by using a
waitpid(). However, this conflicts with the event loop whenever a child is
waited for since the event loop waits for all the pids. As a consequence,
we cannot be sure our waitpid() will be the one that will catch the end of
the snapshotter.

Change-Id: I07325f9e9f7eb0d453950819bdcc58740dc372fb
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>