core-macros: enhance NEXTARG macro
This commit enhances the NEXTARG macro by adding an assert; before that
it was possible to call NEXTARG when argc == 0, which led to a bad memory
access.
Plus, a NEXTPTRARG macro is added, which allows to call the macro without
checking argc first, and returns NULL when there is no more argument.

Change-Id: I5e304da1a106ddf7d383bc470096f3832bb6c425
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>