iop-rpc-channel: fix regression on invalid encoding handling
2615af55a026eae0d3a5af0d2c929e6828669ef4 forced a connection close for
every kind of "invalid encoding". But a constraint violation is also an
"invalid encoding" and we don't want to close the connection for
a miss-usage of a SDK for example. We also don't want to emit a warning
in this case.

In addition, in the case we do want to emit a warning (protocol invalid
encoding) we shouldn't use "logger_is_traced()".

Change-Id: Ia56318967070bf8380b807cf34a93a10926988ce