http: avoid deadlock in tests
We used the blocking API to write our data in the HTTP tests. However when
the TCP reading buffer wasn't large enough (which was the case on centos5)
we blocked in the write() call.

So we fix the test by using a non-blocking API and we also write more data
to easily reproduce this behavior on more recent linux.

Change-Id: Ica07eb6d54500d12da3f439182b7166e01a1b9e8