el: have an informative error in case of inotify overflow.
We previously had an assert so, two possibilites: an overflow is
encountered in dev mode, in that case this creates some support :), or we
are in release mode and the error is simply ignored. When overflow get
ignored, some files may be definitively missed by the event loop, causing
possible out-of-sync issues in the product.

Refs: #37451
Change-Id: I5364c36981582d292e0ca2034b5fb4c82c47b5a0
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>