www/coverage: just pick any intersec.so
The coverage is launching behave in both the .build and the root
directory.
As a result the productdir is ./.build-coverage-centos-6i/one.

Issue:
Excluding '*/\.*' from the find will just always prove to be empty so
intersec.so can't be loaded and we can't access to the web page.

Fix:
Instead of excluding the hidden directories we just want to use any of the
intersec.so we find as they are all the same anyway.

Change-Id: I5bfc13989b2d76a5cd603533acbb13795a17eea3