core-types: assert OPT_ISSET in OPT_VAL
Ensure no access is done to a non-set optional value.
This patch also introduces a new helper OPT_VAL_P which is needed in the
platform and mmsx updates: at many places in the repo, a pointer to the
optional value was used, which was not possible anymore with the assert.

Change-Id: I08ec752a294731f057968321e63d0ee0a361e8c2