iop: do not return an error when fetching unset optional field value.
Currently, when fetching a field value and if the field value is not set,
an error is returned.

This behavior is limited, since sometimes, we need to know that the value
is not set.

A fix is to add an extra boolean variable to tell is the value is set, and
do not return an error in the case of an unset optional value.

Refs: #38257
Change-Id: I4f5e7568147683a570019282fb58515add5641d0