lib-common/ci: backtrace on coredump
During check target, we'll now detect new cores. For each of them, we will
extract backtrace and append it to report : it will be more easier to
detect coredump on buildbot.

Script will also try to detect first frame where intersec code is used to
display locals variables and arguments of function.

Note: script was added in tools

Change-Id: I02eb4dba11a529c853549d28643d04b61279b405