iop-mib: usage used by all "product"2mib exe.
All executables, that will enhance the creation of a MIB, have to use the
usage defined in this patch.

N.B: When a new MIB needs to be created, for example for a new product, a
"product"2mib.c file has to be written on the same shape than
platform2mib.c.

Change-Id: Id02d50911207497c56beae5be0f3788f9186cd3f