iop-mib: refactoring, dealing with types.
This patch enhance a better management of snmp objects types.

To write "Integer32 (1..2147483647)" is only necessary in the case of a
table index. In all of the other cases it is not necessary to give the
limit, writing "Integer32" is enough.

Moreover in the SEQUENCE, that lists all of the entries of a table, when
an entry of a table has a type enum it should only be written INTEGER.

This patch fix those two points.

Change-Id: I7c88e7bb9228b59ea5b7f13359aef617f0ec7fa2