Revert "iop-dso: allow to force pkg fixup when loading DSO"
This reverts commit 6daadc6afaa4637e0c06b3ac8f1557c32d974b7c.
This patch was dangerous because it allowed to patch a DSO which was build
with our build system, and thus was const.

This triggered a segfault on centos 7 nughtly builds:
https://git.corp/buildbot/builders/nightly-rh7/builds/25/steps/test/logs/stdio

I've tried to fix this issue by reworking constness in IOP (cf.
https://git.corp/r/#/c/46274/), but in fact this feature is finally not
used anymore in One...

This also partially reverts ddd74d182, which introduced a test requiring
to fixup a const DSO.

Change-Id: Id31f1de2f300747c94c0098806f1f7b944389b54