iop-mib: one unique function useful during MIB generation.
Thanks to this patch, the function iop_mib() now generates entirely a MIB.
This function gathers the argument parsing, the writing of the MIB buffer,
and the writing of the MIB file using the MIB buffer.

When creating a new "product"2mib.c main function, it will be enough
defining the IOP packages that will be used to create the MIB, defining
the different revisions of the MIB and calling the function iop_mib().

Change-Id: I3a8e9b62f618c18493a339b4697f0a0b4ef70212