iop-json: Introduce @include in the unpacker.
This commit introduces the @include syntax in the unpacker.

This is only supported when unpacking a file (so when using
t_iop_junpack_file or t_iop_junpack_ptr_file).

In this first version, it is only supported for strings (comes for
struct/unions in a next patch).

Unit tests are included.

Change-Id: Ice2c7b88aabec2b71f94df67f3b18f440557cf80
Refs: #38556