unix: make eventfd() available to all.
Previously, we had a eventfd() implemented in the thrjobs in order to
provide an efficient and safe way to wake the main loop from the other
threads. This was hidden there because we supported platforms that may not
have the eventfd feature in the kernel. However, nowadays, we only support
centos 6+ which ave the eventfd, so we can expose the feature for other
usages.

Change-Id: Ice72e51926f8466ced1b5aa30d1964ada4d15336
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>