iop-mib: fix tests when not run in a standalone lib-common.
The tests are expected to be run from a directory where the path
"lib-common/test-data/snmp/mibs/", which may be true when run from
platform or lib-common, but not when run on mmsx.

As a consequence, the nightly tests were failing.

Rewrite a bit the z_check_wanted_file() helper to make use of the
z_cmddfd_g directory fd which points to the directory that contains the
zchk binary being run.

Also make that function a real zchk-helper to avoid panics in its middle,
use Z_ASSERT instead.

Change-Id: I07fb0c2c0b4ed7df08ad4876dc0654ece05fed2a
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>