iopc/snmp: check each tbl has an index.
When declaring a snmpTbl in IOP, it is now needed to tag at least one of
its field with the attribute '@snmpIndex'.

The tag '@snmpIndex' specifies that the field will be used as an index the
table.

This required to add the flag --check-snmp-table-has-index in order to not
break the compilation of older IOP files.

Change-Id: Ic27da7813194d2c414fb426595347a6c8f9affdc