iop-mib: dealing with multiple indexes
Previously:
-----------

For each snmp tables one generic index was created. But, it appeared that
some products may want to declare several indexes.

This patch:
-----------

If some of the 'snmpTbl' object are marked using the attribute @snmpIndex
then they will be used as index for the table. This patch introduces this
feature.

Change-Id: I1819b907af0ba6f7db51e3e22c227e377b9328a1