core-module: add helper to manually force the shutdown of the module.
A C module has been implementing to wrap the platform bigdata query API in
a Java program thanks to the Java Native Interface.

When the Java Virtual Machine is stopping a segfault is triggered. The
function 'mem_stack_pool_wipe()' is called on an unitialized 't_pool_g'
object.

The t_pool_g object is initialized (and wiped) using the thr_hooks
mechanism. The crash is then due to the fact that the start and the
shutdown are done in separate threads.

This is an extract of the associated backtrace from GDB:

(gdb) thread 1
0  0x00007f93a9b4d165 in *__GI_raise (sig=<optimized out>)
1  0x00007f93a9b503e0 in *__GI_abort
2  0x00007f93a9551d49 in os::abort(bool)
3  0x00007f93a96adb4f in VMError::report_and_die()
4  0x00007f93a9558dcf in JVM_handle_linux_signal
5  <signal handler called>
6  0x00007f939b23b3fa in mem_stack_pool_wipe
7  0x00007f939b23b6af in t_pool_wipe
8  0x00007f939b29fc5f in thr_hooks_shutdown
9  0x00007f939b23ed56 in module_shutdown
10 0x00007f939b23ebba in notify_shutdown
11 0x00007f939b23ee80 in module_shutdown
12 0x00007f939b23eff7 in module_release
13 0x00007f939b23f1e4 in module_hard_shutdown
14 0x00007f939b23f2f2 in _module_shutdown
15 0x00007f93aa6ef4cc in ?? () from /lib64/ld-linux-x86-64.so.2
16 0x00007f93a9b51ae2 in __run_exit_handlers
17 0x00007f93a9b51b35 in *__GI_exit
18 0x00007f93a93818c7 in vm_direct_exit(int)
19 0x00007f93a96b332a in VM_Operation::evaluate()
20 0x00007f93a96b1e41 in VMThread::evaluate_operation(VM_Operation*)
21 0x00007f93a96b20da in VMThread::loop()
22 0x00007f93a96b2652 in VMThread::run()
23 0x00007f93a9550c02 in java_start(Thread*)
24 0x00007f93aa2b4b50 in start_thread (arg=<optimized out>)
25 0x00007f93a9bf695d in clone
26 0x0000000000000000 in ??

(gdb) thread 7
[...]

(gdb) bt
0  0x00007f93aa2b5ee5 in pthread_join
1  0x000000000040698c in ?? ()
2  0x00000000004022d6 in ?? ()
3  0x00007f93a9b39ead in __libc_start_main
4  0x00000000004030f1 in ?? ()
5  0x00007fff96e62708 in ?? ()
6  0x000000000000001c in ?? ()
[...]

To fix it, the destructor should be protected and manually called at
shutdown.

Change-Id: Id6a7f08fd552dbed079ce97e36a5b0063b32986f
Refs: #38496