lib-common: writev, fix comparaison with IOV_MAX
The writev function takes a buffer and its length as argument. This length
should be inferior or equal to IOV_MAX.

The file_writev function takes too a buffer and its length as argument,
checks the length and then call writev with buffer length + 1 as argument.

So the file_writev buffer length argument should be strictly inferior to
IOV_MAX.

This check was wrong: we ensured that the buffer length + 1 was strictly
inferior to IOV_MAX.

It was annoying when calling file_writev inside a loop and managing the
IOV_MAX constraint. It leaded to write:
MIN(len, IOV_MAX - 2)

Change-Id: I6b398a663a93e2bf4c6e60350c99a15bd0b3f6f7