iop/printf: unify formats '%*pS' and '%*pC'
Only '%*pS' format is kept and the macro IOP_OBJ_FMT_ARG_FLAGS() is
reworked so it generates suitable arguments for '%*pS'.

Note: in order to avoid a double evaluation of the object given as a
parameter in IOP_OBJ_FMT_ARG(), the struct 'iop_struct_value' is extended
so the desc can be null for objects.

Change-Id: I414f7e96602abb57ba2833f714386e437ecf29b8