iopc-dso: Rework iopc_dso_build error handling.
This commit adds a output error sb to iopc_dso_build instead of displaying
the error on the console. This will allow the caller to add it some
context, and return it as a RPC error for example.

Also stop calling e_fatal when failing to call gcc; this could be annoying
when we'll compile DSOs at runtime (and not only at startup).

Change-Id: I609033a9d0413fa7fe91dba7ac0948cca5c449cd
Refs: #38559