iop-json: Fix @include skipping.
When unpacking a class, the unpacker first looks for the '_class' field.
It runs on all the fields until it is found.

This was buggy in the case a @include was present before the '_class'
field, because this case was forgotten in skip_val.

Fix that, and add a test.

Change-Id: I75a67862c54f8470c5e564eaab7591551e4aa5e6