iopc: Do not call _exit on error.
Problem:
~~~~~~~~
Now that we want to use the iopc library at runtime when updating the
bigdata schema (cf. #38559), we just can't call _exit and terminate the
program on error.
Instead, we have to provide a real error handling, and also clean memory
on errors, which wasn't done at all...

Solution:
~~~~~~~~~
To write this commit, I've just modified the 'fatal' and 'fatal_loc'
macros to not call _exit, and return -1 instead. Then I had to make it
compile, propagate the error at the highest level, and dealing with memory
in case of error.
It seems simple, but trust me, it was REALLY painful :-(.

All the failing iopc tests are now passing in asan without calling _exit,
which makes me think there should not remain a lot of memleaks in this
code...

Change-Id: I32981857472095e64418bc6f67e87d9af8f63022
Refs: #38559