iopc/iopc-dso: Allow to print only file relative part in case of error.
Problem:
~~~~~~~~
When using the iop-dso library to build the event.iop embedded in the
bigdata schema (cf. #38559), an event.iop file is created in a temporary
directory. This file is then compiled using iopc_dso_build, and in case of
error the output sb_t is given back to the RPC caller.

The problem is that the returned errors looked like:
 "/tmp/event-iop-3hyrlG/event.iop:3:1: error: ..."

Which is really not something we want to show to the clients.

Solution:
~~~~~~~~~
This commit allows to specify the prefix directory of the IOP file to
compile when using iopc_dso_build. In case of error, this prefix can be
hidden (or not, depending on the arguments).

This way, the errors look like:
 "event.iop:3:1: error: ..."

Change-Id: Ie2e72eb86559078c8aca4f147d712577839a1865
Refs: #38559