mem[stack/ring]: avoid to exceed MEM_ALLOC_MAX
When the memory pool reached an average allocation size of 16MB we tried
to allocate a memory block exceeding MEM_ALLOC_MAX and thus we aborted.

It's a nonsense to preallocate so much memory. So we clip the
preallocation size to a decent value. I choose here a value of 100MB
because I don't want a too big difference with the current behavior (which
is, on the paper, unlimited). A really sane value should be something like
10MB.

Closes: #39120
Change-Id: I844c689bc98f488119f60914d1f32c506eb838a3