iopc: Fix 'same as' in interfaces when RPCs are reordered.
Problem:
~~~~~~~~
In the case where two RPCs of the same interface had the same arguments
and were not ordered in the same order as their tags, the generated C code
was buggy because RPC with the lowest tag was dumped before, but was using
the arguments of the other one (dumped after), leading to a forward
declaration.
This is because the 'same as' of the interfaces are evaluated in the order
of the RPCs, and not of their tags.

This issue was observed when building https://git.corp/r/#/c/54370/

Solution:
~~~~~~~~~
Move the sort by tags of the RPCs in the interfaces from the C dumper to
the typer, where 'same as' are evaluated.

Also add a test, and bump iopc version.

Change-Id: Icaf09f3d42dd2d7e17e6b1dc5bf3cc162fabf01c