iop-rpc-channel: Ensure errno is properly set by ic_read.
The caller of ic_read prints errno if it is not 0 in case of failure.
The problem is that ic_check_msg_hdr does not set errno when it fails,
leading to absurd messages displayed by the caller of ic_read.

To fix that, set errno to 0 before calling ic_check_msg_hdr (and
ic_check_msg_hdr_flags).

Change-Id: I8c3a9644e1969019e27202080c4a9d40819f83ea