ichannels: Add 'is_trusted' flag.
Problem:
~~~~~~~~
When a daemon receives a big RPC query for a RPC he does not implement,
the message was rejected by ic_check_msg_hdr because the RPC is not
implemented.
As a consequence, the ichannel was closed (because this makes ic_read
fail).

This is an issue because we don't have this behavior when the query is
smaller than 10M (we simply reply with an error), and this becomes a real
big issue when this happens between a child and the master : the child
gets disconnected and commits suicide.

Solution:
~~~~~~~~~
Add a 'is_trusted' flag in the ichannels. We won't close the connection
when receiving big requests for unimplemented RPCs from trusted ichannels
(and in that case the error will be replied when all the message is
received).
We'll set the is_trusted flag in platform for proclaimed ichannels.

Change-Id: I77071d6bfda08b73bdeeb1c2ec31a295936aa6e9