iop: Add iop_struct_check_backward_compat.
This function allows to check the backward compatibility (in binary and/or
in json, depending on the given flags) between two iop_struct_t.

Tests were added, checking iop_struct_check_backward_compat behaves as
expected, and that that the structures are actually backward
compatible/incompatible by packing/unpacking them.

Change-Id: I7e24fbb4a4be7d9ae1449dc2846d5a275ac71806
Refs: #38559