iop_obj_(dyn)vcast: check the casted pointer is not const
There are two versions of each function: one casting to a non-const
pointer (vcast), another one casting to a const pointer (ccast).

When casting a const pointer, only the second one should be used, as the
constness of the pointer should not be casted out.

This was not verified, but is now. This will imply several updates in the
upper repositories.

Change-Id: I6ad05c8a912eedfc939851e0eecdf15e4f0b53c8