iop: Fix boolean normalization in packer.
This tentative of normalization actually had no effet, because gcc
considers that a bool is already normalized.

According to the C norm, it is indeed supposed to be normalized, but
it happens when assigning a boolean with an uninitialized value, like
in https://git.corp/r/#/c/54992/1/platform/bigdata/db-event.blk.

Anyway, this patch fixes the normalization by using a type which is not
already normalized, and compatible with bool.

Note that this was seen following 294a2edcbda305b6, which makes some mmsx
tests fail.

Change-Id: I4c95da3e0aa7752b31a0eeafb154a08f3242d4e7