improve the usefulness of the expect() macro
expect() allows to act like an assert() in debug releases and like
a simple if() in production. In my opinion expect() is some kind of soft
e_panic(), we want to detect an abnormal behavior but we don't want to
abort() because we are in a too critical code-path to be allowed to
abort(). In term of debugging expect() doesn't help much…

Thus I modify expect() to still generate a .debug when the condition fails
but without aborting. This way we'll still got a lot of useful
informations to debug.

Change-Id: I21d7e30ce8cdcd50230a23cd8e3776895b604c33