iop-mib: Simplify/fix tests.
The main object of this commit is to simplify and fix the
z_check_wanted_file helper, which checked that every line present in the
wanted file was also present in the generated MIB. This was
over-complicated, and did not guarantee at all the generated MIB was
correct.

Change that to simply test they are equal.

Also avoid giving sb_t by value (and use SB_8k).

Change-Id: I4f78a344587cd4a96d0072bc786b219ecd447c5a