io-dso: Move clearing of dso->pkg_h.
This patch follows 3810dcbf80d1e9e, and moves this qm_clear to the place
it should really be: in iop_dso_register (just before using it).

This is because in platform schema library, we'll need to access packages
of DSOs that were unregistered (cf. Idb86ffa06b3c7d).

Change-Id: Ibf69c9703396e7b555f7c7e090519acfd852d91f
Refs: #38559