Build: first stub for building a better web build system.
The build system is now structured in a way that is much more similar to
the backend way. However, it imposes a proper directory structure.

www/
 |- Makefile
 `- modules
     |- mod1
     |   |- ...
     |   |- node_modules
     |   `- htdocs
     `- mod2
         |- ...
         |- node_modules
         `- htdocs

The "node_modules" folder contains the code for the front-end scripts
(javascript or typescript). Note that it is not placed in htdocs anymore,
htdocs will only contains the bundle generated from the content of the
'node_modules' directory.

The Makefile now has the following structure:

 # declare the modules for which you need to
 # build something
 _WWWMODULES=mod1 mod2

 # Declare the script bundles to build for mod1
 mod1_WWWSCRIPTS=blah

 # Declare the sources used to build mod1.js,
 # paths are relative to the scripts/ directory
 # of the module.
 blah_SOURCES = blah.ts toto.js

Change-Id: Iecd6c370a202808216c025d743453f7632e4f4f3
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>