core-mem: supervise our main memory consumers
The aim of this patch is to provide some hints of where to look when
investigating for a memory leak.

We count the memory used by our custom memory pools and we also keep track
of them to easily check their state (would be great to add a mandatory
name when creating a pool).

We also account the memory of our qvectors/qhashes/str-buf because they
are intensively used and generally good candidates of high memory
consumption. However it would be too complex to track their memory when
they used a custom pool, so it will only work when they are used with the
default memory pool (libc).

Change-Id: Id9507457929a4a341c872fc63763ec21bfc401db