Fix utf8_strnlen when string contains invalid UTF8.
In that case, the function was looping endlessly, because utf8_charlen
returns 0 when the given string is invalid UTF8.

Plus utf8_charlen (which only takes a pointer on the string without its
len) should not have been used by utf8_strnlen because it is unsafe.
Fix it so that it takes the length.

utf8_strnlen (and utf8_strlen) can now return a negative value in case of
error, which should be handled by the callers.

Change-Id: Id17c4b1435d875baae5072cbf005bab8d0384e8f