iop-macros: introduce IOP_UNION_CASE_V
The problem is that in some cases, the value contained by the union is not
needed and we just want to know its type (eg. in zchk-iop.c). There,
having to declare a variable that contains something we don't want to use
just damages the readability (and may create naming conflicts).

The solution is to declare a new macro IOP_UNION_CASE_V that will behave
just like IOP_UNION_CASE and IOP_UNION_CASE_P but without the last
argument so that the only information we have is the type, not the value.

The two existing macros are also updated in order to get a compilation
error when IOP_UNION_CASE/_P is used instead of IOP_UNION_CASE_V
(ie. the value is unused).

Change-Id: Idb8fcb6d8d138d658115d80d15e0717530329dde