Build: fix www-coverage
To build www-coverage there is a big hack. The whole repository is copied
into .build-* directory. After that the product is compiled in this
directory. And the previous patch 73ede3a about dirty version, generates
temporary file in .build-* directory with c extension which is moved
somewhere in this case. To avoid to conflict with www-coverage target
we suffix the temporary file with 'c+' in build dir.

Change-Id: Ib96a9d5b9c2d3e5071e477a498651dd57ef9fa39