Build: fix possible race conditions with java compilation
The "javac" command handles the dependencies itself, ie: if a.java
depends on b.java, "javac a.java" will generate a.class and b.class.

The issue is that if "javac a.java" and "javac b.java" are run in
parallel, the "javac a.java" may read b.class while it is not complete,
leading to compilation errors.

Three solutions:
* detect dependencies and compile in the right order: clearly the best
  one, however there do not seem to be a simple way to find those.
* set different output directories for each javac command, then copy back
  the relevant class files.
* force the compilation of java files to not be in parallel.

The second solution was picked. It means the same java file may be
compiled several times (maximum n times, n begin the number of java files
in the produced jar). This is clearly not ideal, but given the small
number of java files we have, and the benefit of running the compilation
in parallel, it was considered the best compromise.

Change-Id: Id019bfa955b0eb136916bab9a83704cae3536701