file-log: fix potential segfault when closing a non opened file
The issue is that when calling log_file_close to destroy a log-file,
log_file_flush is called, and expects the _internal pointers to be
non-NULL.

However, if a log file was created but not opened, those pointers are
NULL, and calling log_file_close is the only way to free the created
structure.

 file = log_file_new();
 if (log_file_open(file, ...) < 0) {
     log_file_close(&file); /* < segfault */
 }

log_file_flush is modified to not segfault if the internal pointers are
NULL. Another solution would be to have log_file_delete and
log_file_close, but that would mean two different functions to destroy the
structure, which is bad.

Change-Id: Ie45b3110ac6f40389a5b31087a55e7638b0a5c1d