iopc[parser]: Use packages names as key in packages hash table.
Problem:
~~~~~~~~
I just faced an issue in the way iopc resolved its paths, when trying to
use a structure from qrrd/iop/qrrdquery.iop in the package
platform/platform/statd.iop.
opc simply failed to compile statd.iop with the following error:

  XXX/qrrd/iop/qrrdquery.iop:252:5: error: unable to find file
  `platform.iop` in the include path

Here is what happened:
* when compiling statd.iop, it was added in the global packages hash table
  with the key 'platform/platform/statd.iop'.
* in its dependencies, the package `platform.statd` has (indirectly
  through the package `platform.cfg`) the package 'platform'. It was
  parsed and added in the global packages hash table with the key
  `platform/platform.iop`.
* in order to resolve the structure from the package qrrdquery, it
  properly found qrrd/iop/qrrdquery.iop (because qrrd/iop is in the
  include path),
* qrrdquery.iop uses platform.Exception, so iopc tried to find a package
  called `platform.iop` in the packages hash table, which is not found
  because its key is `platform/platform.iop`.

This shows that the packages caching system of iopc is fragile because
based on paths.

Solution:
~~~~~~~~~
Consider that the key in this hash table is the name of the packages
instead of their paths (here `platform` instead of
`platform/platform.iop`). This works because packages names are supposed
to be unique in an iopc environment, and solves our problem here.

Change-Id: I28e4acd391b9ed180aea6901ce75f788d57feebc