el: add a el_child_get_status and simplify a bit the code.
Since 7c1728b3de26f4, the el_child_register procedure first check that the
child is still alive and if its not the case, it collects the child and
store its status in order to be able to fire the el_t callback on the next
event loop.

As a consequence, if, for some reason (for example, synchronous wait of a
QPS snapshot), we want to manually perform the waitpid() after
el_child_register(), we must handle the situation where the child may
already has been collected. Since the collection has been done by the el,
the status can only be retrieved from it, so we must expose a way to do
so.

Change-Id: Icacca0636b3e807cd6bf1ce9ac040c46ae394dd2
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>