Config: force dwarf2 version in debug/asan profiles
clang >= 3.5 and gcc >= 4.8 emit dwarf4 debugging symbols by default,
which is not supported by gdb 7.4 in wheezy.

It's supported by gdb 7.5, and debian jessie has 7.7.

For gcc it's not a real problem since gcc 4.8 is not available on wheezy,
but our custom clang 3.7 is, so force -gdwarf-2 by default

On newer systems, dwarf4 can be kept with make P=debug USE_DWARF4=1

Change-Id: I0370b718b1a2a660d2e13cf0ed07fe26e7324bd6