Build[java]: fix copy of class files back to build dir
When compiling a java file, the file is first written in a tmp directory,
before being copied into the build one. This is to avoid conflicts because
compiling a java file may compile other at the same time.

For example, with "A.java" and "B.java", with A depending on B, compiling
A.java gives A.class and B.class.

The compilation works by compiling each file in its own directory:
 javac A.java -o .build/.../tmpdir/A/

The directory contains A.class and B.class

Then copying it back in .build/...
This step was buggy as all files were copied back, not just A.class, so if
B.java was compiled before, the copying would fail.
To solve this, copy only the relevant files, ie
 .build/.../tmpdir/A/A*.class

This solves some compilation errors:
https://git.corp/buildbot/builders/review-platform.centos6+/builds/20197/

Change-Id: I88fe30e1cda75c15c88eb2b1b855bb44c5556dcc