new python logger class
IntersecLogger and IntersecLogHandler were based on e_log instead of
loggers, so define a new python logger class and use it in
IntersecLogHandler

The final goal is to send logs from minion python to the logd using
IntersecLogHandler. To achieve that, we will need to use C loggers with
platform support, in the minion.so python module. So export a few
functions (py_add_log_constants and pylogger_get_class) that will be used
in minion python module, and allow to override common module in
IntersecLogHandler.

Note: MYHANDLER/LOGGER global variables were used in lms/enabler_bonus.py
which no longer exists in master so this can go as well

Change-Id: I7c3c73de33bcf1121c604c82bead492df8dd5674
Refs: #33195