bit-wah: Introduce buckets.
In order to avoid reaching the memory allocation limit for big wahs, we'll
split the _data vector into vectors having a limited size, called buckets.

This commit just introduces the buckets in the wah_t structure. The code
building wahs was not changed, so wahs have only one bucket for now, but
the enumerators support multi-wahs.

Change-Id: I849dc9a7a9fc2bbcb9f989a867342096feb28f6b