str-buf: add replacement character for sb_add_filtered(_out)
The functions sb_add_filtered* can be used to filter out (or in) the
characters of a given character set. However, the resulting string is then
often hard to read and to understand.

eg: "geom_ss:41:étude!" becomes "geom_ss41tude" if we want it to be
compliant with a platform.Identifier.

New functions are written where a parameter is added to be able to give a
character that is to be added in place of any substrings removed.
Old functions are deprecated but kept to avoid breaking the API. They will
be removed in the future.

"geom_ss:41:étude!" becomes "geom_ss_41_tude_".

Change-Id: I2e19a3d5419fad45478841cff9364e83bdb55065