iop: proper error handling in iop_dso_open
Opening an iop dso with a class id conflict would trigger a fatal error.
Since the iop dso can be changed at runtime (event.iop in platform), we
cannot keep this behavior.

It was not that simple to fix since iop_register_class modifies global
qms, and if opening a dso fails, these global qms must be kept intact. So
I introduced iop_env_t and associated function to work on a local
environment and modify the global environment only if the dso open
succeeds.

Refs: #40555
Change-Id: Ie55c1e8ed02a09ef7c09d5ec79b71cdd65d480b7