lib-common: Python HTTP, stop to flood log with connection info
The error log "no connection ready" floods the syslog. Furthermore, it
was error-prone because this log was emitted either when there is real
connection problem or when we were waiting the server to answer to our
connection request.

AFAICT there is no way to log an error when a connection fails after a
timeout because the socket connection is non blocking and there is no
callback fired when the timeout is reached, the timer is simply
restarted in this case.

Closes: #41082
Change-Id: I3bb24cc6d90a6a6eb1e18c322528f3c2709b1b19