mem: ring, garbage collect the blocks more often
Releasing UINT16_MAX frames is huge, in QMR query, this is reached
only when 3 G events are extracted.

Thus reduce it to 256 which is about 13 M events extracted.

I suppose this patch could have more sides effects that I imagine,
but with it, the memory of the miner is stable.

Change-Id: Ieb578fa96fd2414db1e54a5fb42fff44cf0df9a3