iopc: doxygen, allow a new keyword named example
This new keyword is not a Doxygen documentation standard.

The goal is to be able for RPC to describe inline a example of request
and answer.

And so with the documentation generator be able for RPC exported in
webservice to give the example in XML.

This will help the PS Team a lot.

This first step add the keyword and fill the generated structure.

The syntax is the same that the one for the generic attribute.

Change-Id: I3e18748ad713ffc69ae0a435d0cf7aae92e12594