iopc: preserve formatting of doxygen comments
We want to preserve formatting of iop multiline doxygen comments to use
asciidoc syntax and directly insert it in asciidoc documents.

Previously each paragraph was split in the lexer, then the first paragraph
(or first sentence) was used for details and the rest was merged to form
the details.

Now the brief is put in the first paragraph, the details in the second
paragraph. And no changes for the parameters which are put in different
chunks.
And we preserve all spaces and newlines in brief and details, except at
the beginning and at the end (c.f. tstjson.iop for all the crazy syntaxes
we support)

Change-Id: I7813700ab34579a3c0f5dedbf611db4c2e8e587f
Closes: #40916