httpc: add a connection error callback
The HTTP client uses a non blocking socket to make a connection with a
server and an el_fd to watch activity. In the el_fd event callback, the
socket status is checked. If the system call getsockopt returns an error
different from EINTR and EINPROGRESS the connection is marked as
disconnected and the httpc is deleted.

The error is not thrown to API user.

In order to have useful log and to allow a better connection handling,
this commit adds a callback to httpc_pool_t and httpc_t.

Refs: #41082
Change-Id: I7766a1836bb18e38b447613bef0815f8e46583d6