core-obj: add a 'super' field in object.
That field is a alias to the object with the type of its parent. As a
consequence, we now have a short and type-safe solution to upcast an
object to its parent type (or any of his parent types), simply by
dereferencing 'obj.super'.

The main improvement, is that we don't need using casting anymore for
upcasting an object, casting can be reserved to downcasting (casting to
a child type). Since using casts make it harder to catch programming error
at compile time, being able to avoid them should be helpful.

As a consequence, the generated structure is a bit more complicated than
before, causing some issues with the http libary not compiling anymore
with C++.

Change-Id: I0b42bb8790762f53bbc474506e86d903345addf1
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>