lib-common: handle EVENT_NOACT in socket connect cb
When a connection is set up for a non blocking socket, we use the event
API to watch its activity with an el_fd_t. The event callback is called in
two cases:
- there is an event on the fd,
- a timeout is reached.

For both HTTP and ichannel client we do not check in this callback if the
event which fired it is a no activity event. We call getsockopt and if it
returns no error, considered the connection as set up.

When there is no activity on a socket file descriptor, the socket internal
errno is not set and getsockopt returns no error.

The fix is to simply check if the event is a no activity one and
considered the connection as not connnected in this case.

Minor: return -1 in httpc_on_connect if the connection is not set like it
is done for ichannel.

Refs: #41082
Change-Id: Icef576bf6f2c0cf41813bba26bdaf30e5a0183b5