iopc: fix segfault in error case
In case of an error, we delete both the attributes which are in the local
attrs qv and the one which were added to the package.

So we need to reset the local attrs qv after associating attributes to the
package.

The problem was introduced with 9040c7bd10b929ab5269091bb81ee40dac6b02f9

Change-Id: Id38e5eaa1c3793bd1f809f397349df44e87a4fc2
Closes: #40538