iopc: init a v5 version
The example field cannot be added in the iop_help_t structure
without breaking the compilation of old branches, because iop-internal.h
is defined per branch.

This could only be handled by creating a new iopc version.

In a near future, we will want this field to be accessible in the dso,
in order to generate the documentation more easily. Thus we won't simply
escape this field.  And because having a TODO for dumping the field
once the v5 will be created, we be certainly forgotten.
The v5 is created now (but not used yet), and the field is dump only in
this case.

Change-Id: Ida27eea56aef88b0754f309cbe627d5a8e614f8b