iopc: generate class id for ctype
iopc ctypes are useful when the type fullname is long. For each class, a
constant is defined for its id in the form <class_fullname>__class_id.
This constant is used in iop class macro.

This commit makes iopc to generate this constant for the class with a
ctype too, allowing to use them in iop class macro.

Change-Id: Ief46ce0aa77e6c5036d0d66bf38a7b2f545e8651