iopc: generate a 'super' field in IOP class C-structures.
This allows easier and statically checkable upcasting of IOP classes.
The new field is named super, which is BTW a reserved keyword of the
IOP, so we are sure it won't conflict with the name of a field of the
class. Moreover, this brings the IOP classes on par with the core-obj
classes.

Change-Id: I672edeb16b3022c36ffe8084d5c20dae7b92337d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>