lstr: add lstr_utf8_is_ilike, copy of qdb_strlike
qdb_strlike is used by many products that do not depend on QDB otherwise.
The function is extracted and put in str-l, to avoid the useless
dependency.
To avoid breaking QDB and all products using it, the behavior is kept as
is, even if some helpers in lib-common might be used for some parts of the
code.

Refs: #39253
Change-Id: I82162b8af3d1b995cfceb4e3d6ce68d2935594f1