iop: Do not (un)register parents when (un)registering a class.
In iop_register_class, it seemed natural to also register all the parent
classes (which can be in other packages). This guarantees that all the
chain is registered, which looked safer when I've coded it.

For symmetry, we were also unregistering the parents when unregistering a
class, but THIS is buggy because we can unregister classes of other
packages that we don't want to unregister. This was not a problem before
because all the packages were "static", but could become one now that we
are dynamically loading/unloading event.iop in platform schema.

The consequence of this patch is that all the packages containing classes
must be explicitly registered, while the packages containing only parent
classes could be omitted before. This might cause some subtle regressions
(which is the reason why I submit this patch on master, besides the fact
that this bug has no consequences on older branches) but is definitely
safer.

Change-Id: I507913ffd4694993e12f116949e335c547cc0c5a